Moving c interface to internal package
